tests(core): group instruction lifecycle coverage
